UI tweaks and stop debugger launching immediately on app launch. Debugger is opened via the native 'System' menu

+++ CPCSharp.Core/CPCRunner.cs
@@ -23,7 +23,7 @@ namespace CPCSharp.Core
     public struct CycleCountObservation
     {
         public double ElapsedMilliseconds { get; set; }
-        public long Count { get; set; }
+        public uint Count { get; set; }
     }
 
     public class CPCRunner
